Refacto customscreen
 - extract reponse sending logic in ButtonActions class
 - extract ack sending logic in ButtonActions class
 - extract button definition  in ButtonActions class
 - move ack logic for result array in ResultTable class
 - move response logic for result array in ResultTable class
 - solve issue with checkbox visible for all lines
 - some cleaning in tests
 - add a global explanation as code comment
 - improve process list code
 - extract card mapping logic from ResultTable to a new class TableRowBuilder
 - extract card filtering logic from ResultTable to a new class CardFilter
 - improve read and ack filter logic
