+++ ui/main/src/app/views/customCardList/buttonActions/ButtonActions-Response.spec.ts
@@ -8,19 +8,15 @@
  */
 
 import {CustomScreenDefinition} from '@ofServices/customScreen/model/CustomScreenDefinition';
-import {CustomCardListView} from './CustomCardListView';
 import {CustomScreenService} from '@ofServices/customScreen/CustomScreenService';
-import {OpfabEventStreamServerMock} from '@tests/mocks/opfab-event-stream.server.mock';
-import {OpfabEventStreamService} from '@ofServices/events/OpfabEventStreamService';
-import {OpfabStore} from '@ofStore/OpfabStore';
-import {RealTimeDomainService} from '@ofServices/realTimeDomain/RealTimeDomainService';
-import {TranslationService} from '@ofServices/translation/TranslationService';
-import {TranslationLibMock} from '@tests/mocks/TranslationLib.mock';
-import {ConfigService} from '@ofServices/config/ConfigService';
-import {DateTimeFormatterService} from '@ofServices/dateTimeFormatter/DateTimeFormatterService';
-import {ConfigServerMock} from '@tests/mocks/configServer.mock';
-import {FilteredLightCardsStore} from '@ofStore/lightcards/FilteredLightcardsStore';
-import {getOneLightCard, setEntities, setProcessConfiguration, setUserPerimeter} from '@tests/helpers';
+import {
+    getOneLightCard,
+    mockTranslation,
+    sendLightCards,
+    setEntities,
+    setProcessConfiguration,
+    setUserPerimeter
+} from '@tests/helpers';
 import {RoleEnum} from '@ofServices/entities/model/RoleEnum';
 import {RightEnum} from '@ofServices/perimeters/model/Perimeter';
 import {ComputedPerimeter} from '@ofServices/users/model/UserWithPerimeters';
@@ -33,15 +29,11 @@ import {ReplaySubject, firstValueFrom} from 'rxjs';
 import {CardTemplateGateway} from '@ofServices/templateGateway/CardTemplateGateway';
 import {Message, MessageLevel} from '@ofServices/alerteMessage/model/Message';
 import {AlertMessageService} from '@ofServices/alerteMessage/AlertMessageService';
+import {ButtonActions} from './ButtonActions';
 
-describe('CustomCardListView - Responses', () => {
-    let opfabEventStreamServerMock: OpfabEventStreamServerMock;
-    let filteredLightCardStore: FilteredLightCardsStore;
-
+describe('CustomCardListView - Button actions - Responses', () => {
     beforeAll(() => {
-        TranslationService.setTranslationLib(new TranslationLibMock());
-        ConfigService.setConfigServer(new ConfigServerMock());
-        DateTimeFormatterService.init();
+        mockTranslation();
         NotificationDecision.init();
         CardTemplateGateway.init();
         CardTemplateGateway.initTemplateFunctions();
@@ -49,12 +41,6 @@ describe('CustomCardListView - Responses', () => {
 
     beforeEach(() => {
         CustomScreenService.clearCustomScreenDefinitions();
-        opfabEventStreamServerMock = new OpfabEventStreamServerMock();
-        OpfabEventStreamService.setEventStreamServer(opfabEventStreamServerMock);
-        OpfabStore.reset();
-        RealTimeDomainService.init();
-        RealTimeDomainService.setStartAndEndPeriod(0, new Date().valueOf() + 1000);
-        filteredLightCardStore = OpfabStore.getFilteredLightCardStore();
     });
     const customScreenDefinition = getCustomScreenDefinitionExample();
 
@@ -97,10 +83,9 @@ describe('CustomCardListView - Responses', () => {
 
     describe('When get responses buttons', () => {
         it('should get button list', () => {
-            CustomScreenService.addCustomScreenDefinition(customScreenDefinition);
-            const view = new CustomCardListView('testId');
+            const buttonActions = new ButtonActions(customScreenDefinition);
 
-            const result = view.getResponseButtons();
+            const result = buttonActions.getResponseButtons();
             expect(result).toEqual([
                 {
                     id: 'button1',
@@ -115,21 +100,18 @@ describe('CustomCardListView - Responses', () => {
         it('should get empty button list if no button list defined', () => {
             const noButtonsCustomScreenDefinition = new CustomScreenDefinition();
             noButtonsCustomScreenDefinition.id = 'testId';
+            const buttonActions = new ButtonActions(noButtonsCustomScreenDefinition);
 
-            CustomScreenService.addCustomScreenDefinition(noButtonsCustomScreenDefinition);
-            const view = new CustomCardListView('testId');
-
-            const result = view.getResponseButtons();
+            const result = buttonActions.getResponseButtons();
             expect(result).toEqual([]);
         });
     });
 
     describe('Response card should', () => {
-        let view: CustomCardListView;
+        let buttonActions: ButtonActions;
         let cardServerMock: CardsServerMock;
         beforeEach(async () => {
-            CustomScreenService.addCustomScreenDefinition(customScreenDefinition);
-            view = new CustomCardListView('testId');
+            buttonActions = new ButtonActions(customScreenDefinition);
 
             cardServerMock = new CardsServerMock();
             cardServerMock.setResponseFunctionForPostCard(
@@ -137,10 +119,10 @@ describe('CustomCardListView - Responses', () => {
             );
             CardsService.setCardsServer(cardServerMock);
 
-            configureProcesses();
-            configureEntities();
+            await configureProcesses();
+            await configureEntities();
             await configureUserPerimeter();
-            await sendTwoCards();
+            sendTwoCards();
 
             async function configureUserPerimeter() {
                 await setUserPerimeter({
@@ -154,8 +136,8 @@ describe('CustomCardListView - Responses', () => {
                 });
             }
 
-            function configureEntities() {
-                setEntities([
+            async function configureEntities() {
+                await setEntities([
                     {
                         id: 'entity1',
                         name: 'entity1 name',
@@ -164,7 +146,7 @@ describe('CustomCardListView - Responses', () => {
                 ]);
             }
 
-            function configureProcesses() {
+            async function configureProcesses() {
                 const myState = new State();
                 myState.response = {state: 'myState'};
 
@@ -172,10 +154,10 @@ describe('CustomCardListView - Responses', () => {
                 statesList.set('myState', myState);
 
                 const process = [new Process('myProcess', '1', 'my process label', null, statesList)];
-                setProcessConfiguration(process);
+                await setProcessConfiguration(process);
             }
 
-            async function sendTwoCards() {
+            function sendTwoCards() {
                 const card = getOneLightCard({
                     publisher: 'entity0',
                     publisherType: 'ENTITY',
@@ -193,17 +175,14 @@ describe('CustomCardListView - Responses', () => {
                     entitiesAllowedToRespond: ['entity1'],
                     id: 'id2'
                 });
-                opfabEventStreamServerMock.sendLightCard(card);
-                opfabEventStreamServerMock.sendLightCard(card2);
-
-                await firstValueFrom(filteredLightCardStore.getFilteredLightCards());
+                sendLightCards([card, card2]);
             }
         });
         it('be send', async () => {
             const responseData = new Map<string, any>();
             responseData.set('id1', {});
             responseData.set('id2', {});
-            await view.clickOnButton('button1', responseData);
+            await buttonActions.sendResponsesWhenUserClicksOnResponseButton('button1', responseData);
 
             expect(cardServerMock.cardsPosted[0].process).toBe('myProcess');
             expect(cardServerMock.cardsPosted[0].state).toBe('myState');
@@ -219,7 +198,7 @@ describe('CustomCardListView - Responses', () => {
             const responseData = new Map<string, any>();
             responseData.set('id1', {comment: 'comment1'});
             responseData.set('id2', {comment: 'comment2'});
-            await view.clickOnButton('button1', responseData);
+            await buttonActions.sendResponsesWhenUserClicksOnResponseButton('button1', responseData);
 
             expect(cardServerMock.cardsPosted[0].process).toBe('myProcess');
             expect(cardServerMock.cardsPosted[0].state).toBe('myState');
@@ -230,7 +209,7 @@ describe('CustomCardListView - Responses', () => {
             expect(cardServerMock.cardsPosted[1].data).toEqual({parentCard: 'id2', comment: 'comment2'});
             expect(cardServerMock.cardsPosted[1].parentCardId).toBe('id2');
         });
-        it('should not be sent and it should show alert to user if custom code return an error', async () => {
+        it('not be sent and should show alert to user if custom code return an error', async () => {
             const responseData = new Map<string, any>();
             responseData.set('id1', {comment: 'comment1'});
             responseData.set('id2', {comment: 'comment2'});
@@ -240,7 +219,7 @@ describe('CustomCardListView - Responses', () => {
                 alertSubject.next(Message);
             });
 
-            await view.clickOnButton('button2', responseData);
+            await buttonActions.sendResponsesWhenUserClicksOnResponseButton('button2', responseData);
             const message = await firstValueFrom(alertSubject.asObservable());
             expect(message.message).toEqual('Error test');
             expect(message.level).toEqual(MessageLevel.ERROR);
