+++ ui/main/src/app/views/customCardList/resultTable/ResultTable-Ack.spec.ts
@@ -1,0 +1,96 @@
+/* Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * See AUTHORS.txt
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ * This file is part of the OperatorFabric project.
+ */
+
+import {CustomScreenService} from '@ofServices/customScreen/CustomScreenService';
+import {getOneLightCard, mockTranslation, setEntities, setProcessConfiguration, setUserPerimeter} from '@tests/helpers';
+import {RightEnum} from '@ofServices/perimeters/model/Perimeter';
+import {ComputedPerimeter} from '@ofServices/users/model/UserWithPerimeters';
+import {AcknowledgmentAllowedEnum, Process, State} from '@ofServices/processes/model/Processes';
+import {ResultTable} from './ResultTable';
+import {Card} from 'app/model/Card';
+import {FieldType} from '@ofServices/customScreen/model/CustomScreenDefinition';
+
+describe('CustomCardListView - Result Array - Acknowledgment', () => {
+    const emptyChildCardsList = new Map<string, Array<Card>>();
+    let resultTable: ResultTable;
+
+    beforeAll(async () => {
+        CustomScreenService.clearCustomScreenDefinitions();
+        mockTranslation();
+        await setEntities([
+            {
+                id: 'entity1',
+                name: 'entity1 name'
+            }
+        ]);
+        await setUserPerimeter({
+            computedPerimeters: [new ComputedPerimeter('myProcess', 'myState', RightEnum.ReceiveAndWrite)],
+            userData: {
+                login: 'test',
+                firstName: 'firstName',
+                lastName: 'lastName',
+                entities: ['entity1']
+            }
+        });
+        const myState = new State();
+        myState.response = {state: 'myState'};
+
+        const stateWithNoAcknowledgment = new State();
+        stateWithNoAcknowledgment.acknowledgmentAllowed = AcknowledgmentAllowedEnum.NEVER;
+
+        const statesList = new Map();
+        statesList.set('myState', myState);
+        statesList.set('myStateWithNoAcknowledgment', stateWithNoAcknowledgment);
+
+        const process = [new Process('myProcess', '1', 'my process label', null, statesList)];
+        await setProcessConfiguration(process);
+        resultTable = new ResultTable({
+            id: 'testId',
+            name: 'name',
+            processIds: [],
+            headerFilters: [],
+            results: {
+                columns: [
+                    {
+                        field: 'testField',
+                        headerName: 'Process',
+                        cardField: 'process',
+                        fieldType: FieldType.STRING
+                    }
+                ]
+            },
+            showAcknowledgmentButton: true
+        });
+    });
+
+    it('isAcknowledgmentPossible should be true if user is allowed to acknowledge ', async () => {
+        const card = getOneLightCard({
+            publisher: 'entity0',
+            publisherType: 'ENTITY',
+            process: 'myProcess',
+            state: 'myState',
+            id: 'id1'
+        });
+
+        const dataArray = resultTable.getDataArrayFromCards([card], emptyChildCardsList);
+        expect(dataArray).toEqual([{cardId: 'id1', testField: 'myProcess', isAcknowledgmentPossible: true}]);
+    });
+    it('isAcknowledgmentPossible should be false if user is not allowed to acknowledge', async () => {
+        const card = getOneLightCard({
+            publisher: 'entity0',
+            publisherType: 'ENTITY',
+            process: 'myProcess',
+            state: 'myStateWithNoAcknowledgment',
+            id: 'id1'
+        });
+
+        const dataArray = resultTable.getDataArrayFromCards([card], emptyChildCardsList);
+        expect(dataArray).toEqual([{cardId: 'id1', testField: 'myProcess', isAcknowledgmentPossible: false}]);
+    });
+});
